refactor: change the structure
