+++ app/components/controlBar/logoutButton.tsx
