+++ app/components/friendBar.tsx
