+++ app/page.tsx
@@ -1,13 +1,20 @@
-import { SocketHandler } from '@/components/global/socketHandler';
+import { MainSidebar } from '@/app/components/mainSidebar';
+import { UtilityArea } from '@/app/components/utilityWindow/utilityArea';
+import { SidebarProvider } from '@/components/ui/sidebar';
 
-export default function Page() {
+export default async function Page({
+  searchParams,
+}: {
+  searchParams: Promise<{ [key: string]: string | undefined }>;
+}) {
+  const currentUser = (await searchParams).user ?? '';
+  const currentMode = (await searchParams).mode ?? 'CHAT';
   return (
     <div>
-      {process.env.BACKEND_URL ? (
-        <SocketHandler backendUrl={process.env.BACKEND_URL} />
-      ) : (
-        <div />
-      )}
+      <SidebarProvider>
+        <MainSidebar currentUser={currentUser} currentMode={currentMode} />
+        <UtilityArea currentUser={currentUser} currentMode={currentMode} />
+      </SidebarProvider>
     </div>
   );
 }
